Start work on parsing immediate operands for Data Processing instructions.

+++ include/chipset/arm2/RegisterFile.h
@@ -20,7 +20,7 @@ class RegisterFile
 {
 public:
 
-    enum class RegisterRef : int
+    enum class RegisterRef : uint8_t
     {
         R0 = 0,
         R1,
